Use itertools to make the solution to part 1 functional
Both solutions were written in an imperative style. Part 1 is now
written in a more functional style, using itertools to fold over the
collection of frequency changes.

I tried to do something similar with part 2 but then I realized that
it wouldn't work the way I wanted because I was trying to fold over an
infinite sequence.