Add failed attempt at using actors for day 7 part 2
I tried to use actors with timers to solve part 2 in real-time. It
looks like there's just enough drift in how long it takes for things
to happen that I'm not getting the right answer. Either that or my
method is bad. Either way, I'm going to try something else. This is
here for posterity.